New3/.../WordArray.hs : Use pinned memory for Karatsuba.
Currently using unsafeCoerce on a Data.Primitive ByteArray, followed
by pointer manipulation followed by another unsafeCoerce to get back
to a ByteArray. This is needed for zero cost array slicing needed for
an efficient implementation of the Karatsuba algorithm. Without pinned
memory, the work-in-progress version of Karatsuba segfault because the
GC moves the memory we're relying on.

Pursing this path to see how fast Karatsuba can be. Will also be
implementing and non-pinned version to compare the speed.