benchmarking : Add benchmarks for big and small Integer multiplication.
I dedicate this commit to Peter Miller (pmiller) who died in the early
hours of 2014/07/27 after a long illness. Peter never programmed in
Haskell but we had many discussions in the early 2000s about the problems
with the languages we were using at the time (C and Python for me and C
and C++ for Peter). It was these discussions which led me to look for
alternatives; first Ocaml and then Haskell. I learned so much from Peter
and had so many good times in his company.

Goodbye my friend. #commitforpmiller