bench-integer.hs : Fix BenchG being incorrectly being imported as BenchS.
I noticed that the GMP and Simple results were almost identical and remembered
that in earlier testing Simple was significantly slower. Turned out that the
generated Check.BenchG file was being imported as both BenchG (correct) and as
BenchS (incorrect).

After fixing this the Simple benchmarks return to where I expected them.