New4: Fix `estimateQuotient` function
The old implementation converted the numerator and denominator to `Double`
divided them and then converted the `Double` back into a `Natural`. This
worked fine unless either the numerator or denominator, when converted into
a `Double`, had a value of `Infinity`.

The new implementation works as above, but instead of using `Double` uses
a psuedo floating point representation of (Word, Int) where the `Word` is
the most significant bits of the `Natural` and the `Int` is a bit shift
value such that:

    (word << shift) <= natural < ((word + 1) << shift)