wip: Implement timesNaturalSplit
For the case of multiplying a very large Natural ABCD by a smaller
large Natural EF, multiple AD by EF and CD by EF and then perform a
a `shiftedAdd2`.

At one point `splitNatural` used some `unsafeCoerce` voodoo, to take
the upper half of the larger Natural. This turned out to be dangerous
because the original `WordArray` might be moved by the GC. For now
the solution is to use an explicit explicit copy.